+++ lib/ogr/layer/extensions.rb
@@ -113,7 +113,7 @@ module OGR
       #   from.
       # rubocop:todo Metrics/PerceivedComplexity
       def point_values(with_attributes = {}) # rubocop:todo Metrics/CyclomaticComplexity
-        return [[]] if feature_count.zero?
+        return [] if feature_count.zero?
 
         field_indices = with_attributes.keys.map { |field_name| find_field_index(field_name) }
         values = Array.new(feature_count) { Array.new(2 + with_attributes.size) }